Fix mode detection (#12)
* Missing refactoring of `getLoc()`

* Fix flags for the location of rhombuses
and add a comment to indicate how these flags work

* Add default case of 'mode -1' when there are no rhombuses

* Remove console logs

* Fix: If there is just one rhombus, it doesn't have a mode
This fix is based on a conversation with Elsa De Luca. Just one rhombus (even in the right position) is not enough to determine the mode of a chant. So, I changed the evaluation of the length of the allWithSETiltLoc array to not only see if it is empty (length == 0), but if it has more than one element (length <= 1, no mode).

---------

Co-authored-by: martha-thomae <thomaemartha@gmail.com>